audio_redactor/requirements.txt
# Core dependencies for audio transcription
faster-whisper>=1.0.0
librosa>=0.9.0

# Streamlit UI
streamlit>=1.45.0

# Audio editing
ffmpeg-python>=0.2.0

# Environment variables
python-dotenv>=1.0.0

# Google Generative AI
google-genai

# Development and debugging
setuptools>=65.0.0 